test_health_check added
