+++ ripper/context.py
@@ -1,7 +1,19 @@
+import os
 from datetime import datetime
 from collections import defaultdict
 from typing import List
 from ripper.sockets import SocketManager
+from ripper.common import readfile
+
+
+def get_headers_dict(base_headers: List[str]):
+    """Set headers for the request"""
+    headers_dict = {}
+    for line in base_headers:
+        parts = line.split(':')
+        headers_dict[parts[0]] = parts[1].strip()
+    
+    return headers_dict
 
 
 class Context:
@@ -22,12 +34,16 @@ class Context:
     url: str = None
 
     # Internal vars
-    user_agents: list = None
-    base_headers: list = None
-    headers = None
+    # user_agents: list = None
+    # base_headers: list = None
+    # headers = None
+    user_agents = readfile(os.path.dirname(__file__) + '/useragents.txt')
+    base_headers = readfile(os.path.dirname(__file__) + '/headers.txt')
+    headers = get_headers_dict(base_headers)
 
     # Statistic
-    start_time: datetime = None
+    # start_time: datetime = None
+    start_time: datetime = datetime.now()
     start_ip: str = ''
     packets_sent: int = 0
     connections_success: int = 0
@@ -37,7 +53,8 @@ class Context:
     connections_check_time: int = 0
     errors: List[str] = []
 
-    cpu_count: int = 1
+    # cpu_count: int = 1
+    cpu_count = max(os.cpu_count(), 1)  # to avoid situation when vCPU might be 0
     show_statistics: bool = False
     current_ip = None
     getting_ip_in_progress: bool = False
